requirements.txt
pyaml
yfinance
time
os